chore: splitting more vue into smaller files
